@@ -10,72 +10,71 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://thedaviddias@github.com/thedaviddias/Front-End-Checklist-App.git"
+    "url": "git+https://thedaviddias@github.com/thedaviddias/Front-End-Checklist.git"
   },
   "author": "David Dias (thedaviddias@gmail.com)",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/thedaviddias/Front-End-Checklist-App/issues"
+    "url": "https://github.com/thedaviddias/Front-End-Checklist/issues"
   },
-  "homepage": "https://github.com/thedaviddias/Front-End-Checklist-App#readme",
+  "homepage": "https://github.com/thedaviddias/Front-End-Checklist#readme",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^9.0.0",
-    "babel-loader": "^7.1.4",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-register": "^6.26.0",
-    "browser-sync": "^2.23.6",
+    "@babel/cli": "7.8.4",
+    "@babel/core": "7.9.0",
+    "@babel/polyfill": "7.8.7",
+    "@babel/preset-env": "7.9.5",
+    "@babel/register": "7.9.0",
+    "babel-eslint": "10.1.0",
+    "browser-sync": "2.26.7",
     "chai": "^4.1.2",
-    "critical": "^1.1.0",
-    "del": "4.1.1",
-    "eslint-config-airbnb-base": "^13.0.0",
-    "eslint-config-prettier": "^3.0.0",
-    "eslint-plugin-flowtype": "^2.46.1",
-    "eslint-plugin-import": "^2.9.0",
-    "eslint-plugin-prettier": "^2.6.0",
-    "flow-bin": "^0.77.0",
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^6.0.0",
+    "critical": "1.3.9",
+    "del": "5.1.0",
+    "eslint-config-airbnb-base": "14.1.0",
+    "eslint-config-prettier": "6.10.1",
+    "eslint-plugin-flowtype": "4.7.0",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-prettier": "3.1.3",
+    "flow-bin": "0.123.0",
+    "gulp": "4.0.2",
+    "gulp-autoprefixer": "7.0.1",
     "gulp-cached": "^1.1.1",
-    "gulp-cdnizer": "^2.0.0",
-    "gulp-changed": "^3.2.0",
+    "gulp-cdnizer": "2.2.0",
+    "gulp-changed": "4.0.2",
     "gulp-cssnano": "^2.1.2",
     "gulp-data": "^1.3.1",
-    "gulp-eslint": "^5.0.0",
+    "gulp-eslint": "6.0.0",
     "gulp-group-css-media-queries": "^1.2.2",
     "gulp-html-replace": "^1.6.2",
     "gulp-htmlmin": "^5.0.0",
-    "gulp-if": "^2.0.2",
-    "gulp-imagemin": "^4.1.0",
+    "gulp-if": "3.0.0",
+    "gulp-imagemin": "7.1.0",
     "gulp-istanbul": "^1.1.3",
-    "gulp-json-concat": "^0.1.1",
-    "gulp-mocha": "^6.0.0",
+    "gulp-json-concat": "0.2.0",
+    "gulp-mocha": "7.0.2",
     "gulp-newer": "^1.4.0",
-    "gulp-plumber": "^1.2.0",
+    "gulp-plumber": "1.2.1",
     "gulp-pug": "^4.0.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-sass": "^4.0.1",
-    "gulp-sourcemaps": "^2.6.4",
-    "gulp-stylelint": "^7.0.0",
-    "gulp-util": "^3.0.8",
+    "gulp-rename": "2.0.0",
+    "gulp-sass": "4.0.2",
+    "gulp-sourcemaps": "2.6.5",
+    "gulp-stylelint": "13.0.0",
     "gulp-webpack": "^1.5.0",
-    "html-webpack-plugin": "^3.0.0",
-    "imagemin-webp": "^4.1.0",
-    "jsdom": "^12.0.0",
+    "html-webpack-plugin": "4.2.0",
+    "imagemin-webp": "5.1.0",
+    "jsdom": "16.2.2",
     "jsdom-global": "3.0.2",
-    "modernizr": "^3.6.0",
+    "modernizr": "3.10.0",
     "postscribe": "^2.0.8",
-    "prettier": "^1.11.1",
+    "prettier": "2.0.4",
     "run-sequence": "^2.2.1",
-    "sinon": "^6.1.3",
-    "stylelint": "^9.1.1",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-scss": "^3.1.3",
-    "uglifyjs-webpack-plugin": "^2.0.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^4.16.1",
-    "webpack-stream": "^5.0.0",
-    "yargs": "^12.0.1"
+    "sinon": "9.0.2",
+    "stylelint": "13.3.2",
+    "stylelint-config-standard": "20.0.0",
+    "stylelint-scss": "3.17.0",
+    "uglifyjs-webpack-plugin": "2.2.0",
+    "url-loader": "4.1.0",
+    "webpack": "4.42.1",
+    "webpack-stream": "5.2.1",
+    "yargs": "15.3.1"
   }
 }
